+++ src/App.js
@@ -1,16 +1,19 @@
 import React from 'react';
-import { Routes, Route } from "react-router-dom";
+import { useRoutes } from "react-router-dom";
 import Home from "./pages/Home";
-import Try from './pages/try';
-import Carousels from './pages/carousels';
-
 import './App.css';
 
-const App = () => (
-  <Routes>
-    <Route path="/" element={<Home />} />
-    
-  </Routes>
-);
+
+function App() {
+  const routes = useRoutes([
+    { path: "/", element: <Home /> }
+  ])
+  return (
+    <>
+    {routes}
+    </>
+  )
+}
+
 
 export default App;
