+++ src/pages/Home.js
@@ -1,71 +1,16 @@
 import React from 'react';
-import { Link } from "react-router-dom";
-import data from '../books.json';
 import "./Home.css";
-import { Carousel, Card } from 'antd';
-import Carousel1 from "../images/carousel1.png";
-import Carousel2 from "../images/carousel2.png";
-import Carousel3 from "../images/carousel3.png";
-import Adventure from "../images/adventure.png";
-import Dictionaries from "../images/dictionaries.png";
-import Fantasy from "../images/fantasy.png";
-import Horror from "../images/horror.png";
+import Container from './container'
 
 
 
 
-const carousel = [Carousel1, Carousel2, Carousel3];
-const catCard = [Adventure, Fantasy, Horror, Dictionaries];
-console.log('adadada', data);
+
 
 function Home() {
   return (
     <>
-      <div className="container" style={{ backgroundColor: '#dddddd !important' }}>
-        <Carousel autoplay className="carousel">
-          {carousel.map((e) => {
-            return <img src={e} className="carousel-img" alt="carousel"></img>;
-          })}
-        </Carousel>
-
-        <div className="cards">
-          {data.map((val, key) => {
-            return (
-              <>
-                <Card className="card">
-                  <h1>{val.title}</h1>
-                  <img src={val.thumbnailUrl} alt="Comics Category" className="card-content" />
-                  <br />
-                  <Link to="/categories" state={"Comics"} className="link">
-                    Shop Now
-                  </Link>
-                </Card>
-              </>
-            );
-          })}
-
-          <Card className="card">
-            <h1>Shop By Category</h1>
-            <div className="card-content">
-              {catCard.map((e) => {
-                return (
-                  <img
-                    src={e}
-                    alt="category"
-                    className="card-category"
-                    onClick={() => console.log("beauty")}
-                  />
-                );
-              })}
-              <br />
-              <Link to="/" className="link">
-                Shop All
-              </Link>
-            </div>
-          </Card>
-
-        </div>
-      </div>
+      <Container/>
     </>
   )
 }
